Create task_4
There is a missing entry in the log file. You'll need to account for that by appending it to the log file. You're given the missing entry stored in a variable named missing_entry.

Use the .write() method and the parameter "a" in the open() function. Be sure to replace each ### YOUR CODE HERE ### with your own code before you run the following cell.

After the portion of the code that writes to the file, another with statement uses the .read() method to read the updated file into the text variable and then display it.