Create task_6
Your next goal is to create a with statement in order to write the IP addresses to the text file you created in the previous step.

You'll first open the file using the "w" parameter. Then, you'll write the IP addresses to the file. Be sure to replace each ### YOUR CODE HERE ### with your own code before you run the following cell. Note that the code cell will contain a with statement that writes to a file but does not display information to the screen, so running it will not produce an output.