Create task_7
In this final step, you'll complete the code you've been writing up to this point. You'll add code to read the file containing IP addresses.

Complete a with statement that reads the text file and stores it in a new variable called text.

Afterwards, display the contents of text and run the cell to explore the result.